hopr_crypto_keypair

Module keystore

Structs

CipherparamsJson

Represents the “cipherparams” part of an encrypted JSON keystore.

CryptoJson

Represents the “crypto” part of an encrypted JSON keystore.

EthKeystore

This struct represents the deserialized form of an encrypted JSON keystore based on the Web3 Secret Storage Definition.

PrivateKeys

Enums

KdfType

Types of key derivition functions supported by the Web3 Secret Storage.

KdfparamsType

Defines the various parameters used in the supported KDFs.